requirements.txt
pandas
requests
rdkit-pypi
networkx
